+++ frontend/Assets/Scripts/Battle/Weapons/Heavy/LandmineWeapon.cs
@@ -10,7 +10,7 @@ namespace Battle.Weapons.Heavy {
         public static WeaponDescriptor Descriptor {
             get {
                 return new WeaponDescriptor(
-                    (int) WeaponId.Landmine,
+                    WeaponId.Landmine,
                     The<WeaponIcons>.Get().Landmine
                 );
             }
