+++ frontend/Assets/Scripts/Battle/Weapons/Thrown/HolyGrenadeWeapon.cs
@@ -6,13 +6,10 @@ namespace Battle.Weapons.Thrown {
 
     public class HolyGrenadeWeapon {
 
-        
-
-        
         public static WeaponDescriptor Descriptor {
             get {
                 return new WeaponDescriptor(
-                    (int) WeaponId.HolyGrenade,
+                    WeaponId.HolyGrenade,
                     The<WeaponIcons>.Get().HolyGrenade
                 );
             }
